+++ src/db/migrations/20251221133433_create_withdrawals_table.ts
@@ -1,42 +1,45 @@
 import type { Knex } from "knex";
 
 export async function up(knex: Knex): Promise<void> {
-  await knex.schema.createTable("withdrawals", (table) => {
-    table.uuid("id").primary();
+  const exists = await knex.schema.hasTable("withdrawals");
+  if (!exists) {
+    await knex.schema.createTable("withdrawals", (table) => {
+      table.uuid("id").primary();
 
-    table
-      .uuid("transaction_id")
-      .notNullable()
-      .references("id")
-      .inTable("transactions")
-      .onDelete("CASCADE");
+      table
+        .uuid("transaction_id")
+        .notNullable()
+        .references("id")
+        .inTable("transactions")
+        .onDelete("CASCADE");
 
-    table
-      .uuid("wallet_id")
-      .notNullable()
-      .references("id")
-      .inTable("wallets")
-      .onDelete("CASCADE");
+      table
+        .uuid("wallet_id")
+        .notNullable()
+        .references("id")
+        .inTable("wallets")
+        .onDelete("CASCADE");
 
-     table.bigInteger("amount").notNullable();
+      table.bigInteger("amount").notNullable();
 
-    table
-      .enu("status", ["pending", "paid", "failed"], {
-        useNative: true,
-        enumName: "withdrawal_status",
-      })
-      .notNullable()
-      .defaultTo("pending");
+      table
+        .enu("status", ["pending", "paid", "failed"], {
+          useNative: true,
+          enumName: "withdrawal_status",
+        })
+        .notNullable()
+        .defaultTo("pending");
 
-    table.string("bank_code", 20).notNullable();
-    table.string("account_number", 20).notNullable();
-    table.timestamps(true, true);
-    table.unique(["transaction_id"]);
-    table.index(["wallet_id"]);
-    table.index(["status"]);
-  });
+      table.string("bank_code", 20).notNullable();
+      table.string("account_number", 20).notNullable();
+      table.timestamps(true, true);
+      table.unique(["transaction_id"]);
+      table.index(["wallet_id"]);
+      table.index(["status"]);
+    });
+  }
 }
 
 export async function down(knex: Knex): Promise<void> {
-    await knex.schema.dropTableIfExists("withdrawals");
+  await knex.schema.dropTableIfExists("withdrawals");
 }
